+++ processor/src/main/java/ca/bc/gov/app/matchers/DoingBusinessAsProcessorMatcher.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static java.util.function.Predicate.not;
