+++ processor/src/main/java/ca/bc/gov/app/matchers/IncorporationNumberProcessorMatcher.java
@@ -1,14 +1,14 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static java.util.function.Predicate.not;
 
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
 import ca.bc.gov.app.dto.legacy.ForestClientDto;
-import ca.bc.gov.app.util.ProcessorUtil;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -16,11 +16,11 @@ import reactor.core.publisher.Mono;
 
 @Component
 @Slf4j
-public class SoleProprietorProcessorMatcher implements ProcessorMatcher {
+public class IncorporationNumberProcessorMatcher implements ProcessorMatcher {
 
   private final WebClient legacyClientApi;
 
-  public SoleProprietorProcessorMatcher(
+  public IncorporationNumberProcessorMatcher(
       @Qualifier("legacyClientApi") WebClient legacyClientApi
   ) {
     this.legacyClientApi = legacyClientApi;
@@ -28,18 +28,18 @@ public class SoleProprietorProcessorMatcher implements ProcessorMatcher {
 
   @Override
   public boolean enabled(SubmissionInformationDto submission) {
-    return List.of("USP", "RSP").contains(submission.clientType());
+    return StringUtils.isNotBlank(submission.incorporationNumber());
   }
 
   @Override
   public String name() {
-    return "Sole Proprietor Matcher";
+    return "Incorporation Number Matcher";
   }
 
   @Override
   public Mono<MatcherResult> matches(SubmissionInformationDto submission) {
 
-    log.info("{} :: Validating {}", name(), submission.corporationName());
+    log.info("{} :: Validating {}", name(), submission.incorporationNumber());
 
     return
         legacyClientApi
@@ -47,12 +47,8 @@ public class SoleProprietorProcessorMatcher implements ProcessorMatcher {
             .uri(
                 uriBuilder ->
                     uriBuilder
-                        .path("/api/search/individual")
-                        .queryParam("firstName",
-                            ProcessorUtil.splitName(submission.corporationName())[1])
-                        .queryParam("lastName",
-                            ProcessorUtil.splitName(submission.corporationName())[0])
-                        .queryParam("dob", submission.dateOfBirth())
+                        .path("/api/search/incorporationOrName")
+                        .queryParam("incorporationNumber", submission.incorporationNumber())
                         .build(Map.of())
             )
             .exchangeToFlux(response -> response.bodyToFlux(ForestClientDto.class))
@@ -61,7 +57,9 @@ public class SoleProprietorProcessorMatcher implements ProcessorMatcher {
             .collectList()
             .filter(not(List::isEmpty))
             .map(values ->
-                new MatcherResult("corporationName", String.join(",", values))
+                new MatcherResult("incorporationNumber", String.join(",", values))
             );
   }
+
 }
+
