+++ processor/src/main/java/ca/bc/gov/app/matchers/IndividualProcessorMatcher.java
@@ -1,14 +1,14 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static java.util.function.Predicate.not;
 
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
 import ca.bc.gov.app.dto.legacy.ForestClientDto;
+import ca.bc.gov.app.util.ProcessorUtil;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -16,11 +16,11 @@ import reactor.core.publisher.Mono;
 
 @Component
 @Slf4j
-public class IncorporationNumberProcessorMatcher implements ProcessorMatcher {
+public class IndividualProcessorMatcher implements ProcessorMatcher {
 
   private final WebClient legacyClientApi;
 
-  public IncorporationNumberProcessorMatcher(
+  public IndividualProcessorMatcher(
       @Qualifier("legacyClientApi") WebClient legacyClientApi
   ) {
     this.legacyClientApi = legacyClientApi;
@@ -28,18 +28,18 @@ public class IncorporationNumberProcessorMatcher implements ProcessorMatcher {
 
   @Override
   public boolean enabled(SubmissionInformationDto submission) {
-    return StringUtils.isNotBlank(submission.incorporationNumber());
+    return "I".equalsIgnoreCase(submission.clientType());
   }
 
   @Override
   public String name() {
-    return "Incorporation Number Matcher";
+    return "Individual Matcher";
   }
 
   @Override
   public Mono<MatcherResult> matches(SubmissionInformationDto submission) {
 
-    log.info("{} :: Validating {}", name(), submission.incorporationNumber());
+    log.info("{} :: Validating {}", name(), submission.corporationName());
 
     return
         legacyClientApi
@@ -47,19 +47,21 @@ public class IncorporationNumberProcessorMatcher implements ProcessorMatcher {
             .uri(
                 uriBuilder ->
                     uriBuilder
-                        .path("/api/search/incorporationOrName")
-                        .queryParam("incorporationNumber", submission.incorporationNumber())
+                        .path("/api/search/individual")
+                        .queryParam("firstName",
+                            ProcessorUtil.splitName(submission.corporationName())[1])
+                        .queryParam("lastName",
+                            ProcessorUtil.splitName(submission.corporationName())[0])
+                        .queryParam("dob", submission.dateOfBirth())
                         .build(Map.of())
             )
             .exchangeToFlux(response -> response.bodyToFlux(ForestClientDto.class))
-            .doOnNext(entity -> log.info("Found a match {}", entity))
             .map(ForestClientDto::clientNumber)
             .collectList()
             .filter(not(List::isEmpty))
             .map(values ->
-                new MatcherResult("incorporationNumber", String.join(",", values))
+                new MatcherResult("corporationName", String.join(",", values))
             );
   }
 
 }
-
