+++ processor/src/main/java/ca/bc/gov/app/matchers/LegalNameProcessorMatcher.java
@@ -1,11 +1,10 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static java.util.function.Predicate.not;
 
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
 import ca.bc.gov.app.dto.legacy.ForestClientDto;
-import ca.bc.gov.app.util.ProcessorUtil;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
@@ -16,11 +15,11 @@ import reactor.core.publisher.Mono;
 
 @Component
 @Slf4j
-public class IndividualProcessorMatcher implements ProcessorMatcher {
+public class LegalNameProcessorMatcher implements ProcessorMatcher {
 
   private final WebClient legacyClientApi;
 
-  public IndividualProcessorMatcher(
+  public LegalNameProcessorMatcher(
       @Qualifier("legacyClientApi") WebClient legacyClientApi
   ) {
     this.legacyClientApi = legacyClientApi;
@@ -28,12 +27,12 @@ public class IndividualProcessorMatcher implements ProcessorMatcher {
 
   @Override
   public boolean enabled(SubmissionInformationDto submission) {
-    return "I".equalsIgnoreCase(submission.clientType());
+    return List.of("I", "USP", "RSP").contains(submission.clientType());
   }
 
   @Override
   public String name() {
-    return "Individual Matcher";
+    return "Legal Name Fuzzy Matcher";
   }
 
   @Override
@@ -47,16 +46,13 @@ public class IndividualProcessorMatcher implements ProcessorMatcher {
             .uri(
                 uriBuilder ->
                     uriBuilder
-                        .path("/api/search/individual")
-                        .queryParam("firstName",
-                            ProcessorUtil.splitName(submission.corporationName())[1])
-                        .queryParam("lastName",
-                            ProcessorUtil.splitName(submission.corporationName())[0])
-                        .queryParam("dob", submission.dateOfBirth())
+                        .path("/api/search/match")
+                        .queryParam("companyName", submission.corporationName())
                         .build(Map.of())
             )
             .exchangeToFlux(response -> response.bodyToFlux(ForestClientDto.class))
             .map(ForestClientDto::clientNumber)
+            .doOnNext(entity -> log.info("Found a match {}", entity))
             .collectList()
             .filter(not(List::isEmpty))
             .map(values ->
