+++ processor/src/main/java/ca/bc/gov/app/matchers/SoleProprietorProcessorMatcher.java
@@ -1,10 +1,11 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static java.util.function.Predicate.not;
 
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
 import ca.bc.gov.app.dto.legacy.ForestClientDto;
+import ca.bc.gov.app.util.ProcessorUtil;
 import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
@@ -15,11 +16,11 @@ import reactor.core.publisher.Mono;
 
 @Component
 @Slf4j
-public class LegalNameProcessorMatcher implements ProcessorMatcher {
+public class SoleProprietorProcessorMatcher implements ProcessorMatcher {
 
   private final WebClient legacyClientApi;
 
-  public LegalNameProcessorMatcher(
+  public SoleProprietorProcessorMatcher(
       @Qualifier("legacyClientApi") WebClient legacyClientApi
   ) {
     this.legacyClientApi = legacyClientApi;
@@ -27,12 +28,12 @@ public class LegalNameProcessorMatcher implements ProcessorMatcher {
 
   @Override
   public boolean enabled(SubmissionInformationDto submission) {
-    return List.of("I", "USP", "RSP").contains(submission.clientType());
+    return List.of("USP", "RSP").contains(submission.clientType());
   }
 
   @Override
   public String name() {
-    return "Legal Name Fuzzy Matcher";
+    return "Sole Proprietor Matcher";
   }
 
   @Override
@@ -46,18 +47,21 @@ public class LegalNameProcessorMatcher implements ProcessorMatcher {
             .uri(
                 uriBuilder ->
                     uriBuilder
-                        .path("/api/search/match")
-                        .queryParam("companyName", submission.corporationName())
+                        .path("/api/search/individual")
+                        .queryParam("firstName",
+                            ProcessorUtil.splitName(submission.corporationName())[1])
+                        .queryParam("lastName",
+                            ProcessorUtil.splitName(submission.corporationName())[0])
+                        .queryParam("dob", submission.dateOfBirth())
                         .build(Map.of())
             )
             .exchangeToFlux(response -> response.bodyToFlux(ForestClientDto.class))
-            .map(ForestClientDto::clientNumber)
             .doOnNext(entity -> log.info("Found a match {}", entity))
+            .map(ForestClientDto::clientNumber)
             .collectList()
             .filter(not(List::isEmpty))
             .map(values ->
                 new MatcherResult("corporationName", String.join(",", values))
             );
   }
-
 }
