+++ processor/src/test/java/ca/bc/gov/app/matchers/SoleProprietorProcessorMatcherTest.java
@@ -1,13 +1,15 @@
-package ca.bc.gov.app.service.processor;
+package ca.bc.gov.app.matchers;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.okJson;
+import static com.github.tomakehurst.wiremock.client.WireMock.okForContentType;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
+import ca.bc.gov.app.matchers.ProcessorMatcher;
+import ca.bc.gov.app.matchers.SoleProprietorProcessorMatcher;
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
 import java.time.LocalDate;
 import java.util.stream.Stream;
@@ -20,24 +22,24 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.test.StepVerifier;
 
-@DisplayName("Unit Test | Individual Matcher")
-class IndividualProcessorMatcherTest {
+@DisplayName("Unit Test | Sole Proprietor Matcher")
+class SoleProprietorProcessorMatcherTest {
 
   @RegisterExtension
   static WireMockExtension wireMockExtension = WireMockExtension
       .newInstance()
-      .options(wireMockConfig().port(10012))
+      .options(wireMockConfig().port(10010))
       .configureStaticDsl(true)
       .build();
-  private final ProcessorMatcher matcher = new IndividualProcessorMatcher(
-      WebClient.builder().baseUrl("http://localhost:10012").build()
-  );
 
+  private final ProcessorMatcher matcher = new SoleProprietorProcessorMatcher(
+      WebClient.builder().baseUrl("http://localhost:10010").build()
+  );
 
   @Test
   @DisplayName("Name matching")
   void shouldMatchName() {
-    assertEquals("Individual Matcher", matcher.name());
+    assertEquals("Sole Proprietor Matcher", matcher.name());
   }
 
   @ParameterizedTest
@@ -49,11 +51,16 @@ class IndividualProcessorMatcherTest {
       MatcherResult result,
       String mockData
   ) {
+
     wireMockExtension.resetAll();
     wireMockExtension
         .stubFor(
             get(urlPathEqualTo("/api/search/individual"))
-                .willReturn(okJson(mockData))
+                .willReturn(
+                    okForContentType("application/json", mockData)
+                        .withHeader("Content-Type", "application/json")
+                )
+
         );
 
     StepVerifier.FirstStep<MatcherResult> verifier =
@@ -74,22 +81,22 @@ class IndividualProcessorMatcherTest {
     return
         Stream.of(
             Arguments.of(
-                new SubmissionInformationDto("James Frank", LocalDate.of(1985, 10, 4), null, "Y",
-                    "I"),
+                new SubmissionInformationDto("James Frank", LocalDate.of(2023, 4, 5), null, null,
+                    "USP"),
                 true,
                 null,
                 "[]"
             ),
             Arguments.of(
-                new SubmissionInformationDto("Marco Polo", LocalDate.of(1977, 3, 22), null, "Y",
-                    "I"),
+                new SubmissionInformationDto("Marco Polo", LocalDate.of(2023, 9, 12), null, null,
+                    "RSP"),
                 false,
                 new MatcherResult("corporationName", String.join(",", "00000000")),
                 "[{\"clientNumber\":\"00000000\"}]"
             ),
             Arguments.of(
-                new SubmissionInformationDto("Lucca DeBiaggio", LocalDate.of(1951, 12, 25), null,
-                    "Y", "I"),
+                new SubmissionInformationDto("Lucca DeBiaggio", LocalDate.of(2023, 10, 11), null,
+                    null, "USP"),
                 false,
                 new MatcherResult("corporationName", String.join(",", "00000000", "00000001")),
                 "[{\"clientNumber\":\"00000000\"},{\"clientNumber\":\"00000001\"}]"
