+++ processor/src/test/java/ca/bc/gov/app/service/legacy/LegacyLoadingServiceTest.java
@@ -8,7 +8,7 @@ import ca.bc.gov.app.ApplicationConstant;
 import ca.bc.gov.app.TestConstants;
 import ca.bc.gov.app.dto.MatcherResult;
 import ca.bc.gov.app.dto.SubmissionInformationDto;
-import ca.bc.gov.app.service.processor.ProcessorMatcher;
+import ca.bc.gov.app.matchers.ProcessorMatcher;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
